Review comments, fix reconnect logic.
