+++ tuyaface/__init__.py
@@ -15,130 +15,6 @@ from tuyaface.helper import *
 logger = logging.getLogger(__name__)
 
 
-HEART_BEAT_PING_TIME = 5
-HEART_BEAT_PONG_TIME = 5
-
-class TuyaClient(threading.Thread):
-    def __init__(self, device: dict, on_status: callable):
-        super().__init__()
-        self.connection = None
-        self.device = device
-        self.force_reconnect = False
-        self.last_ping = 0
-        self.last_pong = time.time()
-        self.on_status = on_status
-        self.seq=0
-        # socketpair used to interrupt the worker thread
-        self.socketpair = socket.socketpair()
-        self.socket_lock = threading.Lock()
-        self.stop = threading.Event()
-
-    def _ping(self):
-        """ Send a ping message. """
-        self.last_ping = time.time()
-        try:
-            logger.debug("TuyaClient: PING")
-            replies = list(reply for reply in send_request(self.device, tf.HEART_BEAT, connection=self.connection, seq=self.seq))
-            self.seq += 1
-            if replies:
-                logger.debug("TuyaClient: PONG %s", replies)
-                self._reset_pong()
-        except socket.error:
-            self.force_reconnect = True
-
-    def _reset_pong(self):
-        """ Reset expired counter. """
-        self.last_pong = time.time()
-
-    def _is_connection_stale(self):
-        """ Indicates if connection has expired. """
-        if time.time() - self.last_ping > HEART_BEAT_PING_TIME:
-            self._ping()
-
-        return (time.time() - self.last_pong) > HEART_BEAT_PING_TIME + HEART_BEAT_PONG_TIME
-
-    def _connect(self):
-        self.connection = _connect(self.device)
-        self._reset_pong()
-
-    def _interrupt(self):
-        try:
-            # Write to the socket to interrupt the worker thread
-            self.socketpair[1].send(b"x")
-        except socket.error:
-            # The socketpair may already be closed during shutdown, ignore it
-            pass
-
-
-    def run(self):
-        self.connection = _connect(self.device)
-
-        while not self.stop.is_set():
-            try:
-                with self.socket_lock:
-                    if self.force_reconnect:
-                        logger.warning("TuyaClient: reconnecting")
-                        if self.connection:
-                            self.connection.close()
-                            self.connection = None
-
-                    if self.connection == None:
-                        try:
-                            logger.debug("TuyaClient: connecting")
-                            self._connect()
-                            logger.info("TuyaClient: connected")
-                        except Exception:
-                            logger.Exception()
-
-                    if self.connection:
-                        # poll the socket, as well as the socketpair to allow us to be interrupted
-                        rlist = [self.connection, self.socketpair[0]]
-                        can_read, _, _ = select.select(rlist, [], [], HEART_BEAT_PING_TIME/2)
-                        if self.connection in can_read:
-                            data = self.connection.recv(4096)
-                            for reply in _process_raw_reply(self.device, data):
-                                logger.debug("TuyaClient: Got msg %s", reply)
-                                if self.on_status:
-                                    reply = json.loads(reply["data"])
-                                    self.on_status(reply)
-
-                        if self.socketpair[0] in can_read:
-                            # Clear the socket's buffer
-                            self.socketpair[0].recv(128)
-
-                        if self._is_connection_stale():
-                             self.force_reconnect = True
-
-                if not self.connection:
-                    time.sleep(HEART_BEAT_PING_TIME/2)
-            except Exception:
-                logger.exception("TuyaClient: Unexpected exception:")
-
-    def stop(self):
-        self.stop.set()
-        self._interrupt()
-        self.join()
-
-    def status(self):
-        with self.socket_lock:
-            if self.connection == None:
-                self._connect()
-            try:
-                status(self.device, connection=self.connection, seq=self.seq)
-                self.seq += 1
-            except socket.error:
-                self.force_reconnect = True
-
-    def set_state(self, value: bool, idx: int = 1):
-        with self.socket_lock:
-            if self.connection == None:
-                self._connect()
-            try:
-                set_state(self.device, value, idx, connection=self.connection, seq=self.seq)
-                self.seq += 1
-            except socket.error:
-                self.force_reconnect = True
-
 def _generate_json_data(device_id: str, command: int, data: dict):
 
     """
@@ -146,14 +22,14 @@ def _generate_json_data(device_id: str, command: int, data: dict):
     return: json str
     """
 
-    payload_dict = {        
-    
-        tf.CONTROL: {"devId": "", "uid": "", "t": ""}, 
+    payload_dict = {
+
+        tf.CONTROL: {"devId": "", "uid": "", "t": ""},
         tf.STATUS: {"gwId": "", "devId": ""},
         tf.HEART_BEAT: {},
-        tf.DP_QUERY: {"gwId": "", "devId": "", "uid": "", "t": ""},  
-        tf.CONTROL_NEW: {"devId": "", "uid": "", "t": ""}, 
-        tf.DP_QUERY_NEW: {"devId": "", "uid": "", "t": ""},          
+        tf.DP_QUERY: {"gwId": "", "devId": "", "uid": "", "t": ""},
+        tf.CONTROL_NEW: {"devId": "", "uid": "", "t": ""},
+        tf.DP_QUERY_NEW: {"devId": "", "uid": "", "t": ""},
     }
 
     json_data = {}
@@ -174,7 +50,7 @@ def _generate_json_data(device_id: str, command: int, data: dict):
     if data is not None:
         json_data['dps'] = data
 
-    return json.dumps(json_data)  
+    return json.dumps(json_data)
 
 
 def _generate_payload(device: dict, request_cnt: int, command: int, data: dict=None):
@@ -188,20 +64,20 @@ def _generate_payload(device: dict, request_cnt: int, command: int, data: dict=N
             This is one of the entries from payload_dict
         data: The data to be send.
             This is what will be passed via the 'dps' entry
-    """     
+    """
 
     #TODO: don't overwrite variables
     payload_json = _generate_json_data(
-        device['deviceid'], 
-        command, 
+        device['deviceid'],
+        command,
         data
     ).replace(' ', '').encode('utf-8')
-    
+
     header_payload_hb = b''
     payload_hb = payload_json
 
     if device['protocol'] == '3.1':
-        
+
         if command == tf.CONTROL:
             payload_crypt = aescipher.encrypt(device['localkey'], payload_json)
             preMd5String = b'data=' + payload_crypt + b'||lpv=' +  b'3.1||' + device['localkey']
@@ -212,21 +88,21 @@ def _generate_payload(device: dict, request_cnt: int, command: int, data: dict=N
             header_payload_hb = b'3.1' + hexdigest[8:][:16].encode('latin1')
             payload_hb =  header_payload_hb + payload_crypt
 
-    elif device['protocol'] == '3.3':   
-        
+    elif device['protocol'] == '3.3':
+
         if command != tf.DP_QUERY:
             # add the 3.3 header
             header_payload_hb = b'3.3' +  b"\0\0\0\0\0\0\0\0\0\0\0\0"
 
         payload_crypt = aescipher.encrypt(device['localkey'], payload_json, False)
         payload_hb = header_payload_hb + payload_crypt
-    else:                 
-        raise Exception('Unknown protocol %s.' % (device['protocol']))            
+    else:
+        raise Exception('Unknown protocol %s.' % (device['protocol']))
 
     return _stitch_payload(payload_hb, request_cnt, command)
 
-    
-def _stitch_payload(payload_hb: bytes, request_cnt: int, command: int):    
+
+def _stitch_payload(payload_hb: bytes, request_cnt: int, command: int):
     """
     Joins the payload request parts together
     """
@@ -243,16 +119,16 @@ def _stitch_payload(payload_hb: bytes, request_cnt: int, command: int):
 
     # calc the CRC of everything except where the CRC goes and the suffix
     hex_crc = format(binascii.crc32(buffer_hb[:-8]) & 0xffffffff, '08X')
-    return buffer_hb[:-8] + hex2bytes(hex_crc) + buffer_hb[-4:]   
+    return buffer_hb[:-8] + hex2bytes(hex_crc) + buffer_hb[-4:]
 
 
-def _process_raw_reply(device: dict, raw_reply: bytes):          
+def _process_raw_reply(device: dict, raw_reply: bytes):
     """
     Splits the raw reply(s) into chuncks and decrypts it.
     returns json str or str (error)
     """
 
-    a = BitArray(raw_reply)  
+    a = BitArray(raw_reply)
 
     #TODO: don't overwrite variables
     for s in a.split('0x000055aa', bytealigned=True):
@@ -262,14 +138,12 @@ def _process_raw_reply(device: dict, raw_reply: bytes):
         if device['protocol'] == '3.1':
             
             data = sbytes[20:-8]
-            if sbytes[20:21] == b'{':   
-
+            if sbytes[20:21] == b'{':
                 if not isinstance(data, str):
                     data = data.decode()
                 yield {"cmd": cmd, "data": data}
             elif sbytes[20:23] == b'3.1':
-
-                logger.info('we\'ve got a 3.1 reply, code untested')                   
+                logger.info('we\'ve got a 3.1 reply, code untested')
                 data = data[3:]  # remove version header
                 data = data[16:]  # remove (what I'm guessing, but not confirmed is) 16-bytes of MD5 hexdigest of payload
                 data_decrypt = aescipher.decrypt(device['localkey'], data)
@@ -278,7 +152,6 @@ def _process_raw_reply(device: dict, raw_reply: bytes):
         elif device['protocol'] == '3.3':
 
             if cmd in [tf.STATUS, tf.DP_QUERY, tf.DP_QUERY_NEW]:
-                
                 data = sbytes[20:8+int.from_bytes(sbytes[14:16], byteorder='big')]
                 if cmd == tf.STATUS:
                     data = data[15:]
@@ -286,7 +159,7 @@ def _process_raw_reply(device: dict, raw_reply: bytes):
                 yield {"cmd": cmd, "data": data_decrypt}
             elif cmd in [tf.HEART_BEAT]:
                 yield {"cmd": cmd, "data": None}
-    
+
 
 def _select_reply(replies: list):
     """
@@ -294,15 +167,13 @@ def _select_reply(replies: list):
     returns json str
     """
 
-    filtered_replies = list(filter(lambda x: x["data"] != b'json obj data unvalid' and x["data"] != 'json obj data unvalid', replies))
+    filtered_replies = list(filter(lambda x: x["data"] != 'json obj data unvalid', replies))
     if len(filtered_replies) == 0:
-        return None
+        return '{}'
     return filtered_replies[0]["data"]
 
 
-
-
-def _status(device: dict, cmd: int = tf.DP_QUERY, expect_reply: int = 1, recurse_cnt: int = 0, connection=None, seq=0):
+def _status(device: dict, cmd: int = tf.DP_QUERY, expect_reply: int = 1, recurse_cnt: int = 0, connection=None, seq: int = 0):
     """
     Sends current status request to the tuya device
     returns json str
@@ -310,14 +181,14 @@ def _status(device: dict, cmd: int = tf.DP_QUERY, expect_reply: int = 1, recurse
 
     replies = list(reply for reply in send_request(device, cmd, None, expect_reply, connection=connection, seq=seq))
 
-    reply = _select_reply(replies)   
+    reply = _select_reply(replies)
     if not reply and recurse_cnt < 3:
         # some devices (ie LSC Bulbs) only offer partial status with CONTROL_NEW instead of DP_QUERY
         reply = _status(device, tf.CONTROL_NEW, 2, recurse_cnt + 1, seq=seq)
     return reply
 
 
-def status(device: dict, connection=None, seq=0):
+def status(device: dict, connection=None, seq: int = 0):
     """
     Requests status of the tuya device
     returns dict
@@ -326,12 +197,10 @@ def status(device: dict, connection=None, seq=0):
     #TODO: validate/sanitize request
     reply = _status(device, connection=connection, seq=seq)
     logger.debug("reply: '%s'", reply)
-    if reply == None:
-        return None
     return json.loads(reply)
 
 
-def set_status(device: dict, dps: dict, connection=None, seq=0):
+def set_status(device: dict, dps: dict, connection=None, seq: int = 0):
     """
     Sends status update request to the tuya device
     returns dict
@@ -340,13 +209,13 @@ def set_status(device: dict, dps: dict, connection=None, seq=0):
     #TODO: validate/sanitize request
     tmp = { str(k):v for k,v in dps.items() }
     replies = list(reply for reply in send_request(device, tf.CONTROL, tmp, 2, connection=connection, seq=seq))
-    
+
     reply = _select_reply(replies)
-    logger.debug("reply: %s", reply)       
+    logger.debug("reply: %s", reply)
     return json.loads(reply)
 
 
-def set_state(device: dict, value: bool, idx: int = 1, connection=None, seq=0):
+def set_state(device: dict, value: bool, idx: int = 1, connection=None, seq: int = 0):
     """
     Sends status update request for one dps value to the tuya device
     returns dict
@@ -370,40 +239,41 @@ def _connect(device: dict, timeout:int = 2):
         connection = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
         connection.setsockopt(socket.IPPROTO_TCP, socket.TCP_NODELAY, 1)
         connection.settimeout(timeout)
-        connection.connect((device['ip'], 6668)) 
-        return connection       
+        connection.connect((device['ip'], 6668))
+        return connection
     except Exception as e:
-        logger.warning('Failed to connect to %s. Retry in %d seconds' % (device['ip'], 1))         
-        raise e       
+        logger.warning('Failed to connect to %s. Retry in %d seconds' % (device['ip'], 1))
+        raise e
 
 
-def send_request(device: dict, command: int = tf.DP_QUERY, payload: dict = None, max_receive_cnt: int = 1, connection = None, seq=0):
+def send_request(device: dict, command: int = tf.DP_QUERY, payload: dict = None, max_receive_cnt: int = 1, connection: socket.socket = None, seq: int = 0):
     """
     Connects to the tuya device and sends the request
     returns json str or str (error)
     """
 
+
     if max_receive_cnt <= 0:
-        return        
+        return
 
     if not connection:
-        connection = _connect(device)           
+        connection = _connect(device)
 
     if command >= 0:
         request = _generate_payload(device, seq, command, payload)
         logger.debug("sending command: [%s] payload: [%s]" % (command,payload))
         try:
-            connection.send(request)                  
+            connection.send(request)
         except Exception as e:
             raise e
 
     try:
-        data = connection.recv(4096)  
-            
-        for reply in _process_raw_reply(device, data):            
+        data = connection.recv(4096)
+
+        for reply in _process_raw_reply(device, data):
             yield reply
     except socket.timeout as e:
-        pass    
-    except Exception as e: 
-        raise e    
+        pass
+    except Exception as e:
+        raise e
     yield from send_request(device, -1, None, max_receive_cnt-1, connection, seq)
